requirements.txt
obs_websocket_py==1.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
